io/i2c: remove command argument from Read and Write
I2C devices support command-less transmission. This CL is removing the
command argument from Read and Write methods.

Users still can pass it as the first byte in the buf if the device
requires a command for writing. Reading from a specific register
will be supported in the future.

Also, switching back to read(2) and write(2) calls since the I2C dev
interface document is listing them as an equivalent of the IOCTLs.

Change-Id: Icc236ebac0f92d471ac5f13205fa5f9882fef07d
Reviewed-on: https://go-review.googlesource.com/22510
Reviewed-by: Jaana Burcu Dogan <jbd@google.com>