errors: add support for Is and As methods
An error can implement an Is method to
allow errors to define themselves as being
equal to other errors using the Is function.

An As method can be implemented by errors
for them to pose as another error type using
the As function.

Change-Id: Ib7fcf45acd18d1bca936dbb8c1d075e1a0b62510
Reviewed-on: https://go-review.googlesource.com/c/147877
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>