cmd/apidiff: in -incompatible mode, don't show header
Currently if there's some breaking change we're OK with, we have to do
something like this:

```
apidiff -incompatible | \
    grep -v ThingImOkWith | \
    grep -v "Incompatible changes" | \
    (! read)
```

This CL changes the -incompatible flag to not show the header, meaning we
don't have to grep exclude it out. The experience ends up being a bit nicer.

Also, apart from the scripting niceties we get from this CL, the header is not
really required if I already specified that I want incompatible changes.

Change-Id: I6e40fcb36a14255d85f8cc0c3db7a8cf58a840d2
Reviewed-on: https://go-review.googlesource.com/c/155459
Reviewed-by: Jonathan Amsterdam <jba@google.com>