errors/fmt: don’t indent first line of detail in absence of message
Newlines within regular messages are now also indented
in detail mode.

The colon ‘:’ is still strictly interpreted as a separator of
the regular message and the detail part. This means that
there is no colon after the first line of a detail-only message.

The indenting formatter is now always called in detail mode.
This does not show any differences in the benchmarks
and also simplifies the codes.

Change-Id: I33e948771bf77a27e0e5421072b5eb0d497a4644
Reviewed-on: https://go-review.googlesource.com/c/155818
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>