notary/internal/tlog: change TileReader.Reject to TileReader.SaveTiles
TileReader has a ReadTiles method that returns unverified tiles,
and then the implementation needs some way to distinguish tiles
that validate from ones that don't. The old approach was to let
ReadTiles assume the tiles would validate and have the validator
promise to call Reject for any tiles that did not validate.
This created uncertainty about the state of the tiles immediately
after a ReadTiles call returned - was a Reject on its way or not?

A better approach, taken in this CL, is to have ReadTiles assume
that tiles it returns are unvalidated until the client calls SaveTiles
with the ones that do pass validation.

Addresses a TODO from the initial review of TileReader.

Change-Id: I69c8cf8aecf33b7fa904646258ae5ffbb915a09e
Reviewed-on: https://go-review.googlesource.com/c/exp/+/171858
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Filippo Valsorda <filippo@golang.org>