ebnf: support raw string tokens as used by EBNF in the Go spec
The EBNF in the Go specification uses raw strings for tokens that
contain double quotes or backslashes. This change adds support
for raw string tokens to the ebnf package, so that go_spec.html
is validated successfully.

This change also re-enables the ebnflint test, which validates the
Go spec.

Fixes golang/go#37290