event: switch to kind specific handlers
This removes Handler and adds a handler per kind, and then detects which
kind specific handlers are supported when building an exporter.
This change makes it safe to add new kinds in the future, and is also faster
for kinds that are not handled.

Change-Id: Ic8f57a5da55cfa179d3e98e6e39112b8f87b8303
Reviewed-on: https://go-review.googlesource.com/c/exp/+/318832
Trust: Ian Cottrell <iancottrell@google.com>
Run-TryBot: Ian Cottrell <iancottrell@google.com>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>