vulndb/govulncheck: make extractModules output unique
For large projects, such as k8s, the number of modules different in
memory but with the same import path can be large. This can result in
many db queries for the same import path, which can be a problem if the
db is remote and no caching is turned on. This CL optimizes the loading
by making sure module extraction returns the same module up to import
paths.

This CL also updates the dependencies on vulndb.

Change-Id: I63f09382e318fa33f7e9280c8cc1ddb5694983b7
Reviewed-on: https://go-review.googlesource.com/c/exp/+/342110
Trust: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Roland Shoemaker <roland@golang.org>