slices: avoid mutating elements between length and capacity in Grow
Repeatedly calling Grow on a slice with sufficient capacity should
effectively be a noop. However, this was not the case since it would
zero out the elements between the length and capacity.
Avoid that behavior to ensure Grow remains cheap in the common case.

Also provide a workaround for golang/go#53888 to ensure that the
number of allocations when growing is always 1.

Fixes golang/go#53853

Change-Id: Id52c0d50ae9a97fe8af9acfeb666360859f336c0
Reviewed-on: https://go-review.googlesource.com/c/exp/+/417494
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Eli Bendersky <eliben@google.com>
Reviewed-by: Ian Lance Taylor <iant@google.com>
Run-TryBot: Joseph Tsai <joetsai@digital-static.net>