slog: both AtomicLevel and Level implement Leveler
Introduce the Leveler interface (only method: Level() Level) so that
both AtomicLevel and Level can be used interchangeably.

HandlerOptions.Level is now a Leveler.

The default for HandlerOptions.Level is InfoLevel, meaning
that Debug logs are disabled by default.

This is also now the behavior of the default Handler, which
is used if slog.{Log,Info,Debug,...} are called before
installing a handler. It used to write everything to the
default log.Logger; now it ignores any level below Info.

Change-Id: I0538201cffacd101901179a01f262ae6a5d96797
Reviewed-on: https://go-review.googlesource.com/c/exp/+/431055
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>