slog: add Level.MarshalJSON
Add a MarshalJSON method to Level so that it
marshals as a string.

This lets us remove special cases for Level in the handlers.

It also ensures that others who write handlers that output JSON will
get strings instead of numbers; the numbers are obscure,
uninformative, and could be confused with the values of other
level-numbering schemes.

Also, improve handler tests to check replacement.

Change-Id: Iedc7aa797469922fb9cfcc3b5d804806f3deb5d8
Reviewed-on: https://go-review.googlesource.com/c/exp/+/432183
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>