slog: move runtime.Callers up the stack
runtime.Callers takes time proportional to its skip argument.
So we can make our code faster by calling it with a smaller skip.

Move the call to runtime.Callers closer to the exported functions
of this package. Moving it all the way to the top would
mean a lot of duplicate code, so we compromise and move it partway
up, to a point where we only have to write the call three times.

This change improves our benchmark times from 4 to 20 percent.

Thanks to Austin Clements for cluing me in to this property of
runtime.Callers.

Change-Id: If9dcaa3d6a7f7bc4f60aa2ad86a56ca5b652e62f
Reviewed-on: https://go-review.googlesource.com/c/exp/+/458210
Reviewed-by: Alan Donovan <adonovan@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>