slog: remove "Writing a handler" section.
As I began writing this section, I found myself repeating the rules
and suggestions in the doc for the Handler interface.

I think we should probably continue to put essential information in
the interface doc. A thorough discussion of how to write a handler
is out of scope for package doc.

Change-Id: I0bf8a9efc81b45f1110ccddd52b66988872ae741
Reviewed-on: https://go-review.googlesource.com/c/exp/+/460835
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Alan Donovan <adonovan@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>