slog: change the meaning of slog.bench
Originally I believed that it was possible to greatly reduce the time
to fetch the pc, so I optimistically used the nopc tag for the "default"
slog benchmarks, written to slog.bench.

After talking to Austin Clements, I no longer think the runtime will
be able to significantly improve the time; instead it is up to this
code to speed up pc fetching by reducing the calldepth argument to
runtime.Callers, as was done in https://go.dev/cl/458210.

So this CL writes to slog.bench for the default, untagged code, and
retains the nopc benchmarks in slog-nopc.bench as a lower bound.

I did not re-run the benchmarks; I only renamed the files.
But git does not seem to notice that.

Change-Id: I8e87dd9345db78e0dbd353050b490300e7d65245
Reviewed-on: https://go-review.googlesource.com/c/exp/+/461997
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>