slog: make context an argument of Handle
Pass the Logger's context to the Handler via an argument,
instead of by putting it in a field of the Record.

This makes Handle look more like Enabled.

If it's not problematic to pass the context to the Enabled method
(because it isn't used for cancellation and may be nil), then it
shouldn't be for the Handle method.

Change-Id: I80743acac53926322bbbbe9703ed60d173e597fd
Reviewed-on: https://go-review.googlesource.com/c/exp/+/467797
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Alan Donovan <adonovan@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>