apidiff: fix comparison of named types in other packages
Compare named types from packages other than old and new using their
package-qualified names.

When determining whether two such types corresponded, we were using
the IDs of the types. This was wrong, because the ID for an exported
type is just its name, unqualified.

The full story is a bit more subtle, because when comparing modules,
we want to allow for the module paths to be different.  That is left
as a TODO.

gorelease actually tests for modules with different paths, but its
test had a mistaken assumption (see the README.txt in this CL) so we
had to change it.

Change-Id: Id06eff0b43292088a016d8556e0c7c9c86616289
Reviewed-on: https://go-review.googlesource.com/c/exp/+/513635
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>