cmd/gorelease: migrated `git status --porcelain` to `git diff` The expectation of this application is that it is to be utilized in release pipelines. Due to this expectation, pipelines utilize volume mounting to transfer artifacts (untracked files) across different jobs (specifically referring to GITLAB here).
This is a problem because git status --porcelain checks for untracked files
and returns them as a list. This causes this pipeline to fail unexpectedly.

In testing, `git reset --hard` will not remove these, while a
`git clean -fdx` will correctly remove these files.

I would recommend that in a follow-up commit there is an update to the
available args for this cmd to control if untracked files are to be
included or not (eg: --untracked)