apidiff: avoid comparing instantiated aliases
During the initial step of comparing APIs, we associate
named types in both the old and new worlds. This doesn't work
if the named type is an alias to an instantiated type, so skip
that case. The proper comparison will happen subsequently.

Fixes golang/go#67051.

Change-Id: Ia405f842e8bf5e258cbcc0748a644d2cea83117a
Reviewed-on: https://go-review.googlesource.com/c/exp/+/618617
Reviewed-by: Tim King <taking@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>