slog: make allocation tests less sensitive to compiler improvements
The compiler improved during the Go 1.25 cycle to reduce user
allocations in some cases.

In particular, CL 649079 avoids allocations by better recognition of
literals that are later used in interface conversions. This now causes a
few cases in TestAlloc to fail due to the improvement.

This CL changes TestAlloc to treat the desired number of allocs as a
maximum allowed, rather than a stricter equality check.

In general, there are some trade-offs for whether to check for equality
in allocation tests, but here it seems likely better to be more lenient
toward compiler improvements, including given x/exp/slog is used with
older versions of the Go compiler.

Fixes golang/go#73928

Change-Id: I9643cb5d43d2cc567e616d67c84382fe7279f64f
Reviewed-on: https://go-review.googlesource.com/c/exp/+/677595
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
TryBot-Bypass: Carlos Amedee <carlos@golang.org>